requirements.txt
asgiref==3.9.1
blinker==1.9.0
click==8.1.8
Flask==3.1.2
greenlet==3.2.4
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
pixelmatch==0.3.0
playwright==1.54.0
pyee==13.0.0
python-slugify==8.0.4
text-unidecode==1.3
typing_extensions==4.14.1
Werkzeug==3.1.3
zipp==3.23.0
